Refactor: updated suspended integration cases and fixed excel blank issue
