+++ src/integrationTest/java/uk/gov/hmcts/reform/cwrdapi/CaseWorkerUpdateUserWithFileUploadTest.java
@@ -8,60 +8,82 @@ import uk.gov.hmcts.reform.cwrdapi.util.CaseWorkerConstants;
 
 import java.util.List;
 
+import static com.github.tomakehurst.wiremock.client.WireMock.get;
 import static com.github.tomakehurst.wiremock.client.WireMock.put;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlPathMatching;
 import static java.lang.String.format;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
-public class CaseWorkerUploadForSuspendedUserWithFileUploadTestTest extends FileUploadTest {
+public class CaseWorkerUpdateUserWithFileUploadTest extends FileUploadTest {
 
-    String exceptedSuspenseResponse = "{\"message\":\"Request Completed Successfully\","
-        + "\"message_details\":\"%s record(s) suspended\"}";
 
-    String expectedSuspendFailureResponse = "{\"message\":\"Request completed with partial success. "
+    String exceptedResponseUpdate = "{\"message\":\"Request completed with partial success. "
         + "Some records failed during validation and were ignored.\","
         + "\"message_details\":\"%s record(s) failed validation\","
-        + "\"error_details\":[{\"row_id\":\"%s\",\"error_description\":\"suspend user in UP failed\"}]}";
+        + "\"error_details\":[{\"row_id\":\"%s\",\"error_description\":\"can't modify roles for user in UP\"}]}";
 
     @Test
     public void shouldCreateCaseWorkerUpdateAuditSuccess() throws Exception {
         validateAuditCaseWorkerCreate();
-        modifyUserStatus(200);
-        response = uploadCaseWorkerFile("CaseWorkerUserSuspendedXlsxWithNoPassword.xlsx",
+        List<CaseWorkerAudit> caseWorkerAudits = caseWorkerAuditRepository.findAll();
+
+        String roles = "[\"Senior Tribunal Caseworker\"]";
+        userProfileGetUserWireMock("ACTIVE", roles);
+        modifyUserRoles();
+        response = uploadCaseWorkerFile("CaseWorkerUserUpdateXlsxWithNoPassword.xlsx",
             CaseWorkerConstants.TYPE_XLSX, "200 OK", cwdAdmin);
         String json = getJsonResponse(response);
         assertThat(objectMapper.readValue(json, CaseWorkerFileCreationResponse.class))
-            .isEqualTo(objectMapper.readValue(format(exceptedSuspenseResponse, 1),
+            .isEqualTo(objectMapper.readValue(format(exceptedResponse, 1), CaseWorkerFileCreationResponse.class));
+        List<CaseWorkerAudit> caseWorkerAuditsUpdate = caseWorkerAuditRepository.findAll();
+        assertThat(caseWorkerAuditsUpdate.size()).isEqualTo(2);
+    }
+
+    @Test
+    public void shouldCreateCaseWorkerFailedToUpdateStaleUser() throws Exception {
+        validateAuditCaseWorkerCreate();
+        String roles = "[\"Senior Tribunal Caseworker\"]";
+        userProfileGetUserWireMock("STALE", roles);
+        modifyUserRoles();
+        response = uploadCaseWorkerFile("CaseWorkerUserUpdateXlsxWithNoPassword.xlsx",
+            CaseWorkerConstants.TYPE_XLSX, "200 OK", cwdAdmin);
+        String json = getJsonResponse(response);
+        assertThat(objectMapper.readValue(json, CaseWorkerFileCreationResponse.class))
+            .isEqualTo(objectMapper.readValue(format(exceptedResponseUpdate, 1, 1),
                 CaseWorkerFileCreationResponse.class));
         List<CaseWorkerAudit> caseWorkerAuditsUpdate = caseWorkerAuditRepository.findAll();
         assertThat(caseWorkerAuditsUpdate.size()).isEqualTo(2);
     }
 
     @Test
-    public void shouldCreateCaseWorkerUpdateAuditFailureWithWrongStatus() throws Exception {
+    public void shouldCreateCaseWorkerFailedToUpdateFailedExceptionInFetchingUser() throws Exception {
         validateAuditCaseWorkerCreate();
-        modifyUserStatus(400);
-        response = uploadCaseWorkerFile("CaseWorkerUserSuspendedXlsxWithNoPassword.xlsx",
+        userProfileService.stubFor(get(urlPathMatching("/v1/userprofile.*"))
+            .willReturn(null));
+        modifyUserRoles();
+        response = uploadCaseWorkerFile("CaseWorkerUserUpdateXlsxWithNoPassword.xlsx",
             CaseWorkerConstants.TYPE_XLSX, "200 OK", cwdAdmin);
         String json = getJsonResponse(response);
         assertThat(objectMapper.readValue(json, CaseWorkerFileCreationResponse.class))
-            .isEqualTo(objectMapper.readValue(format(expectedSuspendFailureResponse, 1,1),
+            .isEqualTo(objectMapper.readValue(format(exceptedResponseUpdate, 1, 1),
                 CaseWorkerFileCreationResponse.class));
         List<CaseWorkerAudit> caseWorkerAuditsUpdate = caseWorkerAuditRepository.findAll();
         assertThat(caseWorkerAuditsUpdate.size()).isEqualTo(2);
     }
 
     @Test
-    public void shouldCreateCaseWorkerUpdateAuditExceptionFailure() throws Exception {
+    public void shouldCreateCaseWorkerFailedToUpdateFailedExceptionInModifyingUser() throws Exception {
         validateAuditCaseWorkerCreate();
+        String roles = "[\"Senior Tribunal Caseworker\"]";
+        userProfileGetUserWireMock("ACTIVE", roles);
         userProfileService.stubFor(put(urlPathMatching("/v1/userprofile.*"))
             .willReturn(null));
-        response = uploadCaseWorkerFile("CaseWorkerUserSuspendedXlsxWithNoPassword.xlsx",
+        response = uploadCaseWorkerFile("CaseWorkerUserUpdateXlsxWithNoPassword.xlsx",
             CaseWorkerConstants.TYPE_XLSX, "200 OK", cwdAdmin);
         String json = getJsonResponse(response);
         assertThat(objectMapper.readValue(json, CaseWorkerFileCreationResponse.class))
-            .isEqualTo(objectMapper.readValue(format(expectedSuspendFailureResponse, 1,1),
+            .isEqualTo(objectMapper.readValue(format(exceptedResponseUpdate, 1, 1),
                 CaseWorkerFileCreationResponse.class));
         List<CaseWorkerAudit> caseWorkerAuditsUpdate = caseWorkerAuditRepository.findAll();
         assertThat(caseWorkerAuditsUpdate.size()).isEqualTo(2);
